feed: fix background image exporting

+++ lawnchair/src/ch/deletescape/lawnchair/feed/images/CurrentImageProvider.kt
@@ -6,7 +6,6 @@ import android.database.Cursor
 import android.database.MatrixCursor
 import android.graphics.Bitmap
 import android.net.Uri
-import android.os.CancellationSignal
 import android.os.ParcelFileDescriptor
 import ch.deletescape.lawnchair.feed.images.providers.ImageProvider
 import ch.deletescape.lawnchair.util.extensions.d
@@ -20,6 +19,7 @@ import java.io.FileOutputStream
 class CurrentImageProvider : ContentProvider() {
     companion object {
         val AUTHORITY = "ch.deletescape.lawnchair.feed.FEED_BACKGROUND_IMAGE"
+        val SHARE_QUERIES = mutableListOf<String>()
     }
 
     override fun onCreate(): Boolean {
@@ -27,8 +27,25 @@ class CurrentImageProvider : ContentProvider() {
     }
 
     override fun query(uri: Uri, projection: Array<String>?, selection: String?,
-                       selectionArgs: Array<String>?, sortOrder: String?): Cursor? {
-        val currentBitmap = File(context!!.cacheDir, "feed_background_share.png")
+                       selectionArgs: Array<String>?, sortOrder: String?): Cursor? = synchronized(this) {
+        d("query: query called. lastPathSegment: ${uri.lastPathSegment} shareQueries: $SHARE_QUERIES")
+        val c = MatrixCursor(
+                arrayOf("_id", "_data", "orientation", "mime_type", "datetaken", "_display_name"))
+        c.addRow(arrayOf<Any?>(0, uri.lastPathSegment, 0, "image/png", System.currentTimeMillis(), ""))
+        return c
+    }
+
+    override fun getType(uri: Uri): String? {
+        return "image/png"
+    }
+
+    @Throws(FileNotFoundException::class)
+    override fun openFile(uri: Uri, mode: String): ParcelFileDescriptor? = synchronized(this) {
+        if (!SHARE_QUERIES.contains(uri.lastPathSegment!!)) {
+            return null
+        }
+        d("openFile: openFile called. lastPathSegment: ${uri.lastPathSegment} shareQueries: $SHARE_QUERIES")
+        val currentBitmap = File(context!!.cacheDir, uri.lastPathSegment)
         currentBitmap.delete()
         var flag = false
         GlobalScope.launch {
@@ -42,29 +59,11 @@ class CurrentImageProvider : ContentProvider() {
             flag = true
         }
         while (!flag);
-        val c = MatrixCursor(
-                arrayOf("_id", "_data", "orientation", "mime_type", "datetaken", "_display_name"))
-        c.addRow(arrayOf<Any?>(0, currentBitmap, 0, "image/png", System.currentTimeMillis(),
-                               "Librechair background image"))
-        return c
-    }
-
-    override fun getType(uri: Uri): String? {
-        return "image/png"
-    }
-
-    @Throws(FileNotFoundException::class)
-    override fun openFile(uri: Uri, mode: String): ParcelFileDescriptor? {
-        d("openFile: openFile called")
-        return ParcelFileDescriptor.open(File(context!!.cacheDir, "feed_background_share.png"),
-                                         ParcelFileDescriptor.MODE_READ_ONLY)
-    }
-
-    override fun openFile(uri: Uri, mode: String,
-                          signal: CancellationSignal?): ParcelFileDescriptor? {
-        d("openFile: openFile called")
-        return ParcelFileDescriptor.open(File(context!!.cacheDir, "feed_background_share.png"),
-                                         ParcelFileDescriptor.MODE_READ_ONLY)
+        val descriptor = ParcelFileDescriptor.open(
+                File(context!!.cacheDir, uri.lastPathSegment),
+                ParcelFileDescriptor.MODE_READ_ONLY)
+        d("openFile: descriptor ${uri.lastPathSegment} is $descriptor")
+        return descriptor
     }
 
     override fun insert(uri: Uri, values: ContentValues?): Uri? {
